Working on Vampires
Working on a timed message to send to vampires warning them of a time
change to dawn and to dusk and that their stats have changed. Changed
register for events for BruteListener in both BasicPlugin and the
BruteListener class. Testing a multiple register (implementing object
oriented execution) in BasicPlugin class.